LibAwsChecksums

Documentation for LibAwsChecksums.

LibAwsChecksums.aws_checksums_crc32

LibAwsChecksums.aws_checksums_crc32_ex

LibAwsChecksums.aws_checksums_crc32c

LibAwsChecksums.aws_checksums_crc32c_ex

LibAwsChecksums.aws_checksums_crc64nvme

LibAwsChecksums.aws_checksums_crc64nvme_ex

LibAwsChecksums.aws_checksums_library_clean_up

LibAwsChecksums.aws_checksums_library_init

LibAwsChecksums.aws_checksums_crc32 — Method

aws_checksums_crc32(input, length, previous_crc32)

The entry point function to perform a CRC32 (Ethernet, gzip) computation. Selects a suitable implementation based on hardware capabilities. Pass 0 in the previousCrc32 parameter as an initial value unless continuing to update a running crc in a subsequent call.

Prototype

uint32_t aws_checksums_crc32(const uint8_t *input, int length, uint32_t previous_crc32);

LibAwsChecksums.aws_checksums_crc32_ex — Method

aws_checksums_crc32_ex(input, length, previous_crc32)

The entry point function to perform a CRC32 (Ethernet, gzip) computation. Supports buffer lengths up to size_t max. Selects a suitable implementation based on hardware capabilities. Pass 0 in the previousCrc32 parameter as an initial value unless continuing to update a running crc in a subsequent call.

Prototype

uint32_t aws_checksums_crc32_ex(const uint8_t *input, size_t length, uint32_t previous_crc32);

LibAwsChecksums.aws_checksums_crc32c — Method

aws_checksums_crc32c(input, length, previous_crc32c)

The entry point function to perform a Castagnoli CRC32c (iSCSI) computation. Selects a suitable implementation based on hardware capabilities. Pass 0 in the previousCrc32 parameter as an initial value unless continuing to update a running crc in a subsequent call.

Prototype

uint32_t aws_checksums_crc32c(const uint8_t *input, int length, uint32_t previous_crc32c);

LibAwsChecksums.aws_checksums_crc32c_ex — Method

aws_checksums_crc32c_ex(input, length, previous_crc32c)

The entry point function to perform a Castagnoli CRC32c (iSCSI) computation. Supports buffer lengths up to size_t max. Selects a suitable implementation based on hardware capabilities. Pass 0 in the previousCrc32 parameter as an initial value unless continuing to update a running crc in a subsequent call.

Prototype

uint32_t aws_checksums_crc32c_ex(const uint8_t *input, size_t length, uint32_t previous_crc32c);

LibAwsChecksums.aws_checksums_crc64nvme — Method

aws_checksums_crc64nvme(input, length, previous_crc64)

The entry point function to perform a CRC64-NVME (a.k.a. CRC64-Rocksoft) computation. Selects a suitable implementation based on hardware capabilities. Pass 0 in the previousCrc64 parameter as an initial value unless continuing to update a running crc in a subsequent call. There are many variants of CRC64 algorithms. This CRC64 variant is bit-reflected (based on the non bit-reflected polynomial 0xad93d23594c93659) and inverts the CRC input and output bits.

Prototype

uint64_t aws_checksums_crc64nvme(const uint8_t *input, int length, uint64_t previous_crc64);

LibAwsChecksums.aws_checksums_crc64nvme_ex — Method

aws_checksums_crc64nvme_ex(input, length, previous_crc64)

The entry point function to perform a CRC64-NVME (a.k.a. CRC64-Rocksoft) computation. Supports buffer lengths up to size_t max. Selects a suitable implementation based on hardware capabilities. Pass 0 in the previousCrc64 parameter as an initial value unless continuing to update a running crc in a subsequent call. There are many variants of CRC64 algorithms. This CRC64 variant is bit-reflected (based on the non bit-reflected polynomial 0xad93d23594c93659) and inverts the CRC input and output bits.

Prototype

uint64_t aws_checksums_crc64nvme_ex(const uint8_t *input, size_t length, uint64_t previous_crc64);

LibAwsChecksums.aws_checksums_library_clean_up — Method

aws_checksums_library_clean_up()

Shuts down the internal data structures used by aws-checksums.

Prototype

void aws_checksums_library_clean_up(void);

LibAwsChecksums.aws_checksums_library_init — Method

aws_checksums_library_init(allocator)

Initializes internal data structures used by aws-checksums. MUST be called before using any functionality in aws-checksums. Note: historically aws-checksums lazily initialized stuff and things worked without init. However, DO NOT rely on that behavior and explicitly call init instead.

Prototype

void aws_checksums_library_init(struct aws_allocator *allocator);